feat: Implement vehicle registration component with styling and validation
- Added `RegistroVehiculoComponent` for vehicle registration functionality.
- Created SCSS styles for the vehicle registration dialog, including responsive design.
- Integrated validation for client documents and vehicle details.
- Removed `ValidarAccionMecanicoComponent` and its associated files as part of refactoring.
