+++ src/app/views/shared/components/registro-cliente/registro-cliente.component.scss
