+++ src/app/views/shared/components/registro-vehiculo/registro-vehiculo.component.scss
