--- src/app/views/shared/components/validar-accion-mecanico/validar-accion-mecanico.component.scss
@@ -1,219 +1,0 @@
-:host ::ng-deep {
-  .custom-dialog {
-    .p-dialog {
-      border-radius: 12px;
-      overflow: hidden;
-      box-shadow: 0 10px 30px rgba(0, 0, 0, 0.3);
-    }
-    
-    .p-dialog-header {
-      background: linear-gradient(135deg, #1f295b 0%, #3a4ab0 100%);
-      color: white;
-      padding: 1rem 1.5rem;
-      border-bottom: none;
-      border-radius: 12px 12px 0 0; // ✅ AGREGAR BORDER-RADIUS AL HEADER
-      
-      .p-dialog-title {
-        font-size: 1.1rem;
-        font-weight: 600;
-        margin: 0;
-      }
-      
-      .p-dialog-header-icon {
-        color: white;
-        font-size: 1.2rem;
-        width: 30px;
-        height: 30px;
-        border-radius: 50%;
-        transition: background-color 0.2s;
-        
-        &:hover {
-          background-color: rgba(255, 255, 255, 0.1);
-        }
-        
-        &:focus {
-          background-color: rgba(255, 255, 255, 0.2);
-          outline: none;
-        }
-      }
-    }
-    
-    .p-dialog-content {
-      padding: 0;
-      background-color: white;
-      border-radius: 0 0 12px 12px; // ✅ TAMBIÉN AGREGAR AL CONTENT PARA COMPLETAR
-    }
-  }
-}
-
-.dialog-body-content {
-  padding: 1.5rem;
-  
-  .verification-section {
-    text-align: center;
-    margin-bottom: 1.5rem;
-    
-    .verification-icon {
-      font-size: 3rem;
-      color: #1f295b;
-      margin-bottom: 0.75rem;
-      display: block;
-    }
-    
-    .verification-title {
-      font-size: 1.125rem;
-      font-weight: 600;
-      color: #374151;
-      margin: 0 0 0.5rem 0;
-    }
-    
-    .verification-text {
-      color: #6b7280;
-      font-size: 0.9rem;
-      margin: 0;
-      line-height: 1.4;
-    }
-  }
-  
-  .form-section {
-    margin-bottom: 1.5rem;
-    
-    .pin-label {
-      display: block;
-      margin-bottom: 0.5rem;
-      font-weight: 500;
-      color: #374151;
-      font-size: 0.9rem;
-    }
-    
-    .pin-input {
-      width: 100% !important;
-      padding: 0.75rem;
-      border: 2px solid #d1d5db;
-      border-radius: 8px;
-      font-size: 1rem;
-      text-align: center;
-      letter-spacing: 0.1em;
-      transition: border-color 0.2s, box-shadow 0.2s;
-      
-      &:focus {
-        border-color: #667eea !important;
-        box-shadow: 0 0 0 3px rgba(102, 126, 234, 0.1) !important;
-        outline: none;
-      }
-      
-      &:disabled {
-        background-color: #f9fafb;
-        cursor: not-allowed;
-        opacity: 0.6;
-      }
-      
-      &::placeholder {
-        color: #9ca3af;
-        letter-spacing: normal;
-      }
-    }
-    
-    .error-message {
-      color: #dc2626;
-      font-size: 0.875rem;
-      text-align: center;
-      padding: 0.75rem;
-      background-color: #fef2f2;
-      border: 1px solid #fecaca;
-      border-radius: 6px;
-      margin-top: 0.75rem;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      gap: 0.5rem;
-    }
-    
-    .loading-message {
-      color: #667eea;
-      font-size: 0.875rem;
-      text-align: center;
-      padding: 0.75rem;
-      background-color: #f0f4ff;
-      border: 1px solid #c7d2fe;
-      border-radius: 6px;
-      margin-top: 0.75rem;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      gap: 0.5rem;
-    }
-  }
-  
-  .action-buttons {
-    display: flex;
-    justify-content: flex-end;
-    gap: 0.75rem;
-    padding-top: 1rem;
-    border-top: 1px solid #e5e7eb;
-    
-    .btn {
-      padding: 0.5rem 1rem;
-      border-radius: 6px;
-      font-size: 0.875rem;
-      font-weight: 500;
-      cursor: pointer;
-      border: none;
-      transition: all 0.2s;
-      display: flex;
-      align-items: center;
-      gap: 0.5rem;
-      min-width: 80px;
-      justify-content: center;
-      
-      &:disabled {
-        opacity: 0.6;
-        cursor: not-allowed;
-      }
-      
-      &.btn-secondary {
-        background-color: #6b7280;
-        color: white;
-        
-        &:hover:not(:disabled) {
-          background-color: #4b5563;
-        }
-      }
-      
-      &.btn-primary {
-        background-color: #667eea;
-        color: white;
-        
-        &:hover:not(:disabled) {
-          background-color: #5a67d8;
-        }
-        
-        &:focus {
-          outline: 2px solid #a5b4fc;
-          outline-offset: 2px;
-        }
-      }
-    }
-  }
-}
-
-// Responsive
-@media (max-width: 480px) {
-  :host ::ng-deep .custom-dialog .p-dialog {
-    margin: 1rem;
-    width: calc(100vw - 2rem) !important;
-    max-width: none !important;
-  }
-  
-  .dialog-body-content {
-    padding: 1rem;
-    
-    .action-buttons {
-      flex-direction: column;
-      
-      .btn {
-        width: 100%;
-      }
-    }
-  }
-}
